requirements.txt
google-generativeai
youtube_transcript_api
streamlit
python-dotenv
pathlib